requirements.txt
pyads
pyyaml
termcolor
argparse